+++ QtFresnelZones/drawer.h
@@ -8,7 +8,7 @@ class Drawer
 private:
     Drawer ();
 public:
-    static void drawArrow (QPainter& painter, unsigned x0, unsigned y0, unsigned x1, unsigned y1, double k1 = 15.0, double k2 = 10.0);
+    static void drawArrow (QPainter& painter, double x0, double y0, double x1, double y1, double k1 = 15.0, double k2 = 10.0);
 };
 
 #endif // DRAWER_H
